inject MysqlSupport
temporary fix for now. Refactor it to core later
