+++ bukkit/src/main/java/io/github/wysohn/triggerreactor/bukkit/modules/BukkitDriverModule.java
@@ -36,5 +36,7 @@ public class BukkitDriverModule extends AbstractModule {
         bind(IInventoryHandle.class).to(BukkitInventoryHandle.class);
         bind(IPluginManagement.class).to(BukkitPluginManagement.class);
         bind(TaskSupervisor.class).to(BukkitTaskSupervisor.class);
+
+        bind(MysqlSupport.class);
     }
 }
